backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlmodel
psycopg2-binary
python-multipart
python-dotenv
passlib[bcrypt]
supabase
numpy
open3d
laspy